+++ ImmichServiceCore/immichapi/src/main/java/com/liskovsoft/immichapi/network/dto/AlbumResponseDto.java
@@ -1,0 +1,20 @@
+package com.liskovsoft.immichapi.network.dto;
+
+import com.google.gson.annotations.SerializedName;
+
+public final class AlbumResponseDto {
+    @SerializedName("id")
+    public String id;
+
+    @SerializedName("albumName")
+    public String albumName;
+
+    @SerializedName("assetCount")
+    public int assetCount;
+
+    @SerializedName("albumThumbnailAssetId")
+    public String albumThumbnailAssetId;
+
+    @SerializedName("description")
+    public String description;
+}