+++ ImmichServiceCore/immichserviceinterfaces/src/main/java/com/liskovsoft/immichserviceinterfaces/data/ImmichUser.java
@@ -1,0 +1,9 @@
+package com.liskovsoft.immichserviceinterfaces.data;
+
+public interface ImmichUser {
+    String getId();
+
+    String getName();
+
+    String getEmail();
+}